files added and lines changed

+++ routes/web.php
@@ -29,26 +29,38 @@ use App\Http\Controllers\ListingController;
 Route::get('/',[ListingController::class, 'index']);
 
 //show create form
-Route::get('/listings/create',[ListingController::class, 'create']);
+Route::get('/listings/create',[ListingController::class,
+ 'create'])->middleware('auth');
 
 ///store Listing Data
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 //Show Edit Form
-Route::get('/listings/{listing}/edit',[ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit',[ListingController::class, 'edit'])->middleware('auth');
 
 //Update Listing
-Route::put('/listings/{listing}', [ListingController::class, 'update'] );
+Route::put('/listings/{listing}', [ListingController::class, 'update'] )->middleware('auth');
 
 //Delete Listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy'] );
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'] )->middleware('auth');
 
 
 ///single listings
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 //Show Register/Create form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 //Create new user
 Route::post('/users', [UserController::class, 'store']);
+
+//Log user out
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+//show login form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+//Login user
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
+
+
